+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.10.0] - 2024-10-09
+
 - add import/export plugin
 - open todo in sidebar on click
 
@@ -82,3 +84,41 @@
 ## [0.1.0] - 2023-08-10
 
 - first release
+
+[Unreleased]: https://github.com/powerpaul17/yatoo/compare/v0.10.0...HEAD
+
+[0.10.0]: https://github.com/powerpaul17/yatoo/compare/v0.9.0...v0.10.0
+
+[0.9.0]: https://github.com/powerpaul17/yatoo/compare/v0.8.0...v0.9.0
+
+[0.8.0]: https://github.com/powerpaul17/yatoo/compare/v0.7.0...v0.8.0
+
+[0.7.0]: https://github.com/powerpaul17/yatoo/compare/v0.6.0...v0.7.0
+
+[0.6.0]: https://github.com/powerpaul17/yatoo/compare/v0.5.1...v0.6.0
+
+[0.5.1]: https://github.com/powerpaul17/yatoo/compare/v0.5.0...v0.5.1
+
+[0.5.0]: https://github.com/powerpaul17/yatoo/compare/v0.4.3...v0.5.0
+
+[0.4.3]: https://github.com/powerpaul17/yatoo/compare/v0.4.2...v0.4.3
+
+[0.4.2]: https://github.com/powerpaul17/yatoo/compare/v0.4.1...v0.4.2
+
+[0.4.1]: https://github.com/powerpaul17/yatoo/compare/v0.4.0...v0.4.1
+
+[0.4.0]: https://github.com/powerpaul17/yatoo/compare/v0.3.2...v0.4.0
+
+[0.3.2]: https://github.com/powerpaul17/yatoo/compare/v0.3.1...v0.3.2
+
+[0.3.1]: https://github.com/powerpaul17/yatoo/compare/v0.3.0...v0.3.1
+
+[0.3.0]: https://github.com/powerpaul17/yatoo/compare/v0.2.2...v0.3.0
+
+[0.2.2]: https://github.com/powerpaul17/yatoo/compare/v0.2.1...v0.2.2
+
+[0.2.1]: https://github.com/powerpaul17/yatoo/compare/v0.2.0...v0.2.1
+
+[0.2.0]: https://github.com/powerpaul17/yatoo/compare/v0.1.0...v0.2.0
+
+[0.1.0]: https://github.com/powerpaul17/yatoo/releases/tag/v0.1.0